\name{sixproblems-package}
\alias{sixproblems-package}
\alias{sixproblems}
\docType{package}
\title{
\packageTitle{sixproblems}
}
\description{
\packageDescription{sixproblems}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sixproblems}
\packageIndices{sixproblems}

A collection of trial-level, individual-subject data sets from category
learning experiments that employ the Shepard, Hovland and Jenkins (1961)
abstract stimulus structure. This is a curated collection, meaning that
the package maintainer has verified that each data set accurately
reproduces a stated set of published summaries of those data.

The package is likely to be useful for those wishing to conduct
computational modelling, and/or secondary data analysis, of this
Canonical Independently Replicated Phenomenon (CIRP) in category
learning. It is designed to complement the \code{catlearn} package.

There is a separate dataset (and help file) for each experiment. For a
complete list of included experiments, use \code{library(help =
"sixproblems")}.

 } 
\author{
\packageAuthor{sixproblems}

Maintainer: \packageMaintainer{sixproblems}
}
\references{

  Shepard, R.N., Hovland, C.I., & Jenkins, H.M. (1961). learning and
  memorization of classifications. \emph{Psychological Monographs, 75},
  Whole No. 517.
  
}
\keyword{ package }
\seealso{
  \code{\link[catlearn:catlearn-package]{catlearn}}
}
\examples{

data(nosof94tld)

}
